Implement users default discount rate to order (ADNAPI-158)

+++ src/Adanom.Ecommerce.API.Commands/Commands/Identity/User/Models/UserResponse.cs
@@ -12,7 +12,7 @@
 
         public string PhoneNumber { get; set; } = null!;
 
-        public byte DefaultDiscountRate { get; set; }
+        public decimal DefaultDiscountRate { get; set; }
 
         public bool AllowCommercialEmails { get; set; }
